refactor: restructure code into separate folder
