+++ src/App/TechnicalDashboard/ConnectionHealth/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ConnectionHealthDropdown';
